# src/swagger_files/QuickCode.Demomusti.TrainingModule.Api-Swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "QuickCode - Demomusti - TrainingModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/training-module/training/count": {
      "get": {
        "tags": [
          "Training"
        ],
        "operationId": "TrainingCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-module/training/{id}": {
      "get": {
        "tags": [
          "Training"
        ],
        "operationId": "TrainingGetItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.TrainingDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Training"
        ],
        "operationId": "TrainingUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.TrainingDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.TrainingDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.TrainingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Training"
        ],
        "operationId": "TrainingDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training-module/training/{trainingId}/training-session/{trainingSessionId}": {
      "get": {
        "tags": [
          "Training"
        ],
        "operationId": "TrainingGetTrainingSessionsForTrainingDetails",
        "parameters": [
          {
            "name": "trainingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "trainingSessionsId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "trainingSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.GetTrainingSessionsForTrainingResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training-module/training-sessions/count": {
      "get": {
        "tags": [
          "TrainingSessions"
        ],
        "operationId": "TrainingSessionsCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-module/training-sessions/{id}": {
      "get": {
        "tags": [
          "TrainingSessions"
        ],
        "operationId": "TrainingSessionsGetItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.TrainingSessionDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TrainingSessions"
        ],
        "operationId": "TrainingSessionsUpdate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.TrainingSessionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.TrainingSessionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.TrainingSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TrainingSessions"
        ],
        "operationId": "TrainingSessionsDelete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/training-module/training-sessions/get-by-training/{trainingSessionTrainingId}": {
      "get": {
        "tags": [
          "TrainingSessions"
        ],
        "operationId": "TrainingSessionsGetByTraining",
        "parameters": [
          {
            "name": "trainingSessionTrainingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.GetByTrainingResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.GetTrainingSessionsForTrainingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "trainingId": {
            "type": "integer",
            "format": "int32"
          },
          "sessionName": {
            "type": "string",
            "nullable": true
          },
          "sessionDate": {
            "type": "string",
            "format": "date-time"
          },
          "sessionLocation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QuickCode.Demomusti.TrainingModule.Application.Dtos.Training.TrainingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "courseName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "trainingType": {
            "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Domain.Enums.TrainingType"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/QuickCode.Demomusti.TrainingModule.Domain.Enums.TrainingStatus"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.GetByTrainingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "sessionName": {
            "type": "string",
            "nullable": true
          },
          "sessionDate": {
            "type": "string",
            "format": "date-time"
          },
          "sessionLocation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QuickCode.Demomusti.TrainingModule.Application.Dtos.TrainingSession.TrainingSessionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "trainingId": {
            "type": "integer",
            "format": "int32"
          },
          "sessionName": {
            "type": "string",
            "nullable": true
          },
          "sessionDate": {
            "type": "string",
            "format": "date-time"
          },
          "sessionLocation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QuickCode.Demomusti.TrainingModule.Domain.Enums.TrainingStatus": {
        "enum": [
          "Planned",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "type": "string",
        "x-enumNames": [
          "Planned",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "x-enum-varnames": [
          "Planned",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "x-enumDescriptions": [
          "Training is planned",
          "Training is in progress",
          "Training has been completed",
          "Training has been cancelled"
        ]
      },
      "QuickCode.Demomusti.TrainingModule.Domain.Enums.TrainingType": {
        "enum": [
          "Online",
          "InPerson",
          "Workshop"
        ],
        "type": "string",
        "x-enumNames": [
          "Online",
          "InPerson",
          "Workshop"
        ],
        "x-enum-varnames": [
          "Online",
          "InPerson",
          "Workshop"
        ],
        "x-enumDescriptions": [
          "Online training",
          "In-person training",
          "Workshop training"
        ]
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}